refactored js files, splitting main
